+++ sigrid-ext-app/src/app/models/vs-command-type.ts
@@ -8,4 +8,6 @@ export enum VsCommandType {
   CreateAzureDevOpsWorkItem = 'createAzureDevOpsWorkItem',
   GetAzureDevOpsWorkItemTypes = 'getAzureDevOpsWorkItemTypes',
   FixFindingsWithAi = 'fixFindingsWithAi',
+  CheckSystemOnboarded = 'checkSystemOnboarded',
+  OnboardSystem = 'onboardSystem',
 }